Some changes for delete, update and add methods in the ToDoService; FileCache -> FileCacheService

+++ ToDoList/Services/ToDoService/ToDoService.swift
@@ -8,24 +8,23 @@
 import Foundation
 
 final class ToDoService {
-    let fileCache: FileCache
+    let fileCacheService: FileCacheService
     let networkingService: DefaultNetworkingService
+    var items: [ToDoItem] = []
     private let itemsQueue = DispatchQueue(label: "ToDoService", attributes: [.concurrent])
-    init(fileCache: FileCache, networkingService: DefaultNetworkingService) {
-        self.fileCache = fileCache
+    init(fileCacheService: FileCacheService, networkingService: DefaultNetworkingService) {
+        self.fileCacheService = fileCacheService
         self.networkingService = networkingService
     }
     func getToDoItem(id: String) -> ToDoItem? {
-        return fileCache.get(with: id)
+        return items.first(where: {$0.id == id})
     }
     func getToDoItems(withDone: Bool) -> [ToDoItem] {
-        if withDone {
-            return fileCache.toDoItems
+        return itemsQueue.sync {
+            (!withDone ? items.filter({!$0.done}) : items).sorted(by: {$0.createdAt < $1.createdAt})
         }
-        return fileCache.toDoItems.filter({!$0.done})
     }
     func update(_ item: ToDoItem, queue: DispatchQueue, completion: @escaping (Result<Void, Error>) -> Void) {
-        fileCache.add(item: item)
         networkingService.update(item) { [weak self] result in
             switch result {
             case let .success(item):
@@ -40,7 +39,7 @@ final class ToDoService {
                                          updatedAt: item.updatedAt,
                                          createdAt: item.createdAt,
                                          isDirty: true)
-                self?.fileCache.add(item: dirtyItem)
+                self?.replaceAndSave(item: dirtyItem, queue: queue, completion: completion)
                 queue.async {
                     completion(.failure(error))
                 }
@@ -48,25 +47,35 @@ final class ToDoService {
         }
     }
     func delete(_ id: String, queue: DispatchQueue, completion: @escaping (Result<Void, Error>) -> Void) {
-        fileCache.delete(with: id)
+        guard let index = items.firstIndex(where: {$0.id == id}) else {
+            return
+        }
+        items.remove(at: index)
         networkingService.delete(id) { [weak self] result in
-            print(result)
+            guard let self = self else {return}
             switch result {
             case .success(_):
-                completion(.success(()))
+                self.fileCacheService.saveFile(items: self.items) { _ in
+                    queue.async {
+                        completion(.success(()))
+                    }
+                }
             case let .failure(error):
-                self?.fileCache.addTombstone(tombstone: Tombstone(id: id))
-                completion(.failure(error))
+                self.fileCacheService.addTombstone(tombstone: Tombstone(id: id))
+                queue.async {
+                    completion(.failure(error))
+                }
             }
         }
     }
     func create(_ item: ToDoItem, queue: DispatchQueue, completion: @escaping (Result<Void, Error>) -> Void) {
-        fileCache.add(item: item)
+        items.append(item)
         networkingService.create(item) { [weak self] result in
+            guard let self = self else {return}
             switch result {
             case let .success(item):
-                self?.replaceAndSave(item: item, queue: queue, completion: completion)
-            case let .failure(error):
+                self.replaceAndSave(item: item, queue: queue, completion: completion)
+            case .failure(_):
                 let dirtyItem = ToDoItem(id: item.id,
                                          text: item.text,
                                          importance: item.importance,
@@ -76,113 +85,110 @@ final class ToDoService {
                                          updatedAt: item.updatedAt,
                                          createdAt: item.createdAt,
                                          isDirty: true)
-                self?.fileCache.add(item: dirtyItem)
-                completion(.failure(error))
+                self.replaceAndSave(item: dirtyItem, queue: queue, completion: completion)
             }
         }
     }
     func replaceAndSave(item: ToDoItem, queue: DispatchQueue, completion: @escaping (Result<Void, Error>) -> Void) {
-        replace(item: item) { [weak self] in
-            self?.fileCache.saveFile {result in
-                switch result {
-                case let .failure(error):
-                    queue.async {
-                        completion(.failure(error))
-                    }
-                case .success():
-                    queue.async {
-                        completion(.success(()))
-                    }
+        itemsQueue.async { [weak self] in
+            guard let self = self,
+                  let index = self.items.firstIndex(where: {$0.id == item.id}) else {
+                queue.async {
+                    completion(.success(()))
+                }
+                return
+            }
+            self.items[index] = item
+            self.fileCacheService.saveFile(items: self.items) { result in
+                queue.async {
+                    completion(result)
                 }
             }
         }
     }
-    func merge(items: [ToDoItem], completion: @escaping ([ToDoItem]) -> Void) {
+    func merge(newItems: [ToDoItem], queue: DispatchQueue, completion: @escaping ([ToDoItem]) -> Void) {
         itemsQueue.async { [weak self] in
             guard let self = self else {return}
-            let cacheItems = self.fileCache.toDoItems
-            var newItems: [ToDoItem] = []
-            for item in items {
-                if let cacheItem = cacheItems.first(where: {$0.id == item.id}) {
-                   if let cacheUpdatedAt = cacheItem.updatedAt,
-                      let itemUpdatedAt = item.updatedAt,
-                      cacheUpdatedAt < itemUpdatedAt {
-                        newItems.append(item)
-                   } else {
-                    newItems.append(cacheItem)
-                   }
+            let oldItems = self.items
+            var items: [ToDoItem] = []
+            for newItem in newItems {
+                if let oldItem = oldItems.first(where: {$0.id == newItem.id}) {
+                    if let oldUpdate = oldItem.updatedAt,
+                       let newUpdate = newItem.updatedAt,
+                       oldUpdate <= newUpdate {
+                        items.append(newItem)
+                    } else {
+                        items.append(oldItem)
+                    }
                 } else {
-                    newItems.append(item)
+                    items.append(newItem)
                 }
             }
-            completion(newItems)
+            self.items = items
+            queue.async {
+                completion(items)
+            }
         }
     }
-    func replace(item: ToDoItem, completion: @escaping () -> Void) {
-        itemsQueue.async(flags: [.barrier]) { [weak self] in
-            guard let self = self else {return}
-            self.fileCache.add(item: item)
-        }
-        completion()
-    }
-    func replace(items: [ToDoItem], completion: @escaping () -> Void) {
-        itemsQueue.async(flags: [.barrier]) { [weak self] in
-            guard let self = self else {return}
-            items.forEach({self.fileCache.add(item: $0)})
-        }
-        completion()
-    }
-    func loadFromServer(queue: DispatchQueue, completion: @escaping (Result<[ToDoItem], Error>) -> Void) {
+    func loadData(queue: DispatchQueue, completion: @escaping (Result<[ToDoItem], Error>) -> Void) {
         networkingService.getAll { [weak self] result in
+            guard let self = self else {return}
             switch result {
             case let .success(items):
-                self?.merge(items: items) { items in
-                    queue.async {
-                        self?.fileCache.reloadItems(toDoItems: items) {
+                self.merge(newItems: items, queue: queue) { items in
+                    self.fileCacheService.saveFile(items: items) { _ in
+                        queue.async {
                             completion(.success(items))
                         }
                     }
                 }
             case .failure(_):
-                self?.loadFromFile(queue: queue, completion: completion)
+                self.loadFromFile(queue: queue, completion: completion)
             }
         }
     }
     func loadFromFile(queue: DispatchQueue, completion: @escaping (Result<[ToDoItem], Error>) -> Void) {
-        fileCache.loadFile { result in
+        fileCacheService.loadFile { result in
             queue.async {
                 completion(result)
             }
         }
     }
-    func synchronize(item: ToDoItem?, idToDelete: String?, queue: DispatchQueue,
-                     completion: @escaping (Result<Void, Error>) -> Void) {
-        var dirties = fileCache.getDirties()
-        var tombstones = fileCache.tombstones.map({$0.id})
-        if let item = item {
-            dirties.append(item)
-        }
-        if let id = idToDelete {
-            tombstones.append(id)
-        }
-        networkingService.putAll(addOrUpdateItems: dirties,
-                                 deleteIds: tombstones) { [weak self] result in
-            switch result {
-            case let .failure(error):
-                queue.async {
-                    completion(.failure(error))
-                }
-            case let .success(items):
-                queue.async {
-                    self?.fileCache.clearTombstones()
-                    self?.fileCache.reloadItems(toDoItems: items) {
-                        completion(.success(()))
-                    }
-                }
-            }
-        }
-    }
-    private func needToSynchronize() -> Bool {
-        return fileCache.getDirties().count != 0 || fileCache.tombstones.count != 0
-    }
+//    func synchronize(item: ToDoItem?, idToDelete: String?, queue: DispatchQueue,
+//                     completion: @escaping (Result<Void, Error>) -> Void) {
+//        if let item = item {
+//            let dirtyItem = ToDoItem(id: item.id,
+//                                     text: item.text,
+//                                     importance: item.importance,
+//                                     deadline: item.deadline,
+//                                     color: item.color,
+//                                     done: item.done,
+//                                     updatedAt: item.updatedAt,
+//                                     createdAt: item.createdAt,
+//                                     isDirty: true)
+//            fileCache.add(item: dirtyItem)
+//        }
+//        if let id = idToDelete {
+//            fileCache.addTombstone(tombstone: Tombstone(id: id))
+//        }
+//        networkingService.putAll(addOrUpdateItems: fileCache.getDirties(),
+//                                 deleteIds: fileCache.tombstones.map({$0.id})) { [weak self] result in
+//            switch result {
+//            case let .failure(error):
+//                queue.async {
+//                    completion(.failure(error))
+//                }
+//            case let .success(items):
+//                queue.async {
+//                    self?.fileCache.clearTombstones()
+//                    self?.fileCache.reloadItems(toDoItems: items) {
+//                        completion(.success(()))
+//                    }
+//                }
+//            }
+//        }
+//    }
+//    private func needToSynchronize() -> Bool {
+//        return fileCache.getDirties().count != 0 || fileCache.tombstones.count != 0
+//    }
 }
